factory/templates/fastapi-agentic/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.14
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
